+++ packages/reports/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./artifactHashes";
 export * from "./markdownReport";
 export * from "./portfolioDemo";
+export * from "./sarif";
 export * from "./traceabilityMatrix";